+++ datafusion/proto/tests/cases/roundtrip_logical_plan.rs
@@ -62,6 +62,7 @@ use datafusion_proto::logical_plan::{
     from_proto, DefaultLogicalExtensionCodec, LogicalExtensionCodec,
 };
 use datafusion_proto::protobuf;
+use datafusion_proto_common::protobuf_common;
 
 use prost::Message;
 
@@ -1214,7 +1215,7 @@ fn round_trip_scalar_types() {
 
     for test_case in should_pass.into_iter() {
         let field = Field::new("item", test_case, true);
-        let proto: protobuf::Field = (&field).try_into().unwrap();
+        let proto: protobuf_common::Field = (&field).try_into().unwrap();
         let roundtrip: Field = (&proto).try_into().unwrap();
         assert_eq!(format!("{field:?}"), format!("{roundtrip:?}"));
     }
@@ -1370,7 +1371,7 @@ fn round_trip_datatype() {
     ];
 
     for test_case in test_cases.into_iter() {
-        let proto: protobuf::ArrowType = (&test_case).try_into().unwrap();
+        let proto: protobuf_common::ArrowType = (&test_case).try_into().unwrap();
         let roundtrip: DataType = (&proto).try_into().unwrap();
         assert_eq!(format!("{test_case:?}"), format!("{roundtrip:?}"));
     }
@@ -1394,13 +1395,11 @@ fn roundtrip_dict_id() -> Result<()> {
 
     // encode
     let mut buf: Vec<u8> = vec![];
-    let schema_proto: datafusion_proto::generated::datafusion::Schema =
-        schema.try_into().unwrap();
+    let schema_proto: protobuf_common::Schema = schema.try_into().unwrap();
     schema_proto.encode(&mut buf).unwrap();
 
     // decode
-    let schema_proto =
-        datafusion_proto::generated::datafusion::Schema::decode(buf.as_slice()).unwrap();
+    let schema_proto = protobuf_common::Schema::decode(buf.as_slice()).unwrap();
     let decoded: Schema = (&schema_proto).try_into()?;
 
     // assert
@@ -1450,7 +1449,7 @@ fn roundtrip_field() {
         (String::from("k1"), String::from("v1")),
         (String::from("k2"), String::from("v2")),
     ]));
-    let proto_field: protobuf::Field = (&field).try_into().unwrap();
+    let proto_field: protobuf_common::Field = (&field).try_into().unwrap();
     let returned_field: Field = (&proto_field).try_into().unwrap();
     assert_eq!(field, returned_field);
 }
@@ -1468,7 +1467,7 @@ fn roundtrip_schema() {
             (String::from("k3"), String::from("v3")),
         ]),
     );
-    let proto_schema: protobuf::Schema = (&schema).try_into().unwrap();
+    let proto_schema: protobuf_common::Schema = (&schema).try_into().unwrap();
     let returned_schema: Schema = (&proto_schema).try_into().unwrap();
     assert_eq!(schema, returned_schema);
 }
@@ -1493,12 +1492,12 @@ fn roundtrip_dfschema() {
         ]),
     )
     .unwrap();
-    let proto_dfschema: protobuf::DfSchema = (&dfschema).try_into().unwrap();
+    let proto_dfschema: protobuf_common::DfSchema = (&dfschema).try_into().unwrap();
     let returned_dfschema: DFSchema = (&proto_dfschema).try_into().unwrap();
     assert_eq!(dfschema, returned_dfschema);
 
     let arc_dfschema = Arc::new(dfschema.clone());
-    let proto_dfschema: protobuf::DfSchema = (&arc_dfschema).try_into().unwrap();
+    let proto_dfschema: protobuf_common::DfSchema = (&arc_dfschema).try_into().unwrap();
     let returned_arc_dfschema: DFSchemaRef = proto_dfschema.try_into().unwrap();
     assert_eq!(arc_dfschema, returned_arc_dfschema);
     assert_eq!(dfschema, *returned_arc_dfschema);
